@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, Output } from "@angular/core";
-import { FormBuilder } from "@angular/forms";
+import { UntypedFormBuilder } from "@angular/forms";
 
 @Component({
   selector: 'inline-form',
@@ -20,7 +20,7 @@ export class InlineFormComponent {
     title: [''],
   });
 
-  constructor(private fb: FormBuilder) {}
+  constructor(private fb: UntypedFormBuilder) {}
 
   activeEditing(): void {
     if (this.title) {
